v2/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
Pillow>=10.0.0
PyYAML>=6.0
wandb>=0.16.0
tqdm>=4.65.0
einops>=0.7.0
pywavelets>=1.4.1
timm>=0.9.0
matplotlib>=3.7.0
scipy>=1.11.0
lpips>=0.1.4
